// features/post/src/components/WhatWeDid/WhatWeDid.types.ts
export type ReportProps = {
  hours: number
  activities: number
  tours: number
  offers: number
  days: number
  escrowClosed: string
  homesToured: number
  offersMade: number
  purchaseDate: string
  closingDate: string
  listPrice: string
  purchasePrice: string
  purchaseVsListPrice: string
  closingCredits: string
  milestones: number
  completedActivities: number
  totalHours: number
  totalMileage: string
}